+++ .github/scripts/d_main.d
@@ -1,0 +1,5 @@
+import check : _check;
+
+void main() {
+    _check();
+}